pyproject.toml
[project]
name = "line-agent-secretary"
version = "0.1.0"
description = "LINE AI Agent Secretary System with Google Calendar integration"
requires-python = ">=3.12"
dependencies = [
    # Agent Framework
    "strands-agents>=0.1.0",
    # AWS SDK
    "boto3>=1.34.0",
    # AWS AgentCore
    "bedrock-agentcore>=0.1.0",
    # Google APIs
    "google-api-python-client>=2.116.0",
    "google-auth-httplib2>=0.2.0",
    # LINE Messaging API
    "line-bot-sdk>=3.9.0",
    # Web Framework (for local development)
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "aws-cdk-aws-lambda-python-alpha>=2.221.0a0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

infra = [
    "aws-cdk-lib>=2.120.0",
    "constructs>=10.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "functions", "infra"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
